Video/requirements.txt
absl-py==0.12.0
argcomplete==1.10.0
astor==0.8.1
beautifulsoup4==4.8.0
bs4==0.0.1
cached-property==1.5.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.1.2
Flask-Cors==3.0.10
funcy==1.15
future==0.18.2
gast==0.4.0
gensim==3.8.3
google-pasta==0.2.0
grpcio==1.36.1
h5py==3.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
IMAPClient==2.1.0
importlib-metadata==3.7.3
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.0
moviepy==1.0.3
nltk==3.5
numexpr==2.7.3
numpy==1.20.1
olefile==0.46
opencv-python==4.5.1.48
pandas==1.1.5
pdf2image==1.14.0
pdfminer.six==20181108
Pillow==8.1.2
proglog==0.1.9
protobuf==3.15.6
pycryptodome==3.10.1
pyLDAvis==3.2.2
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.12.0
sklearn==0.0
smart-open==4.2.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
textract==1.6.3
threadpoolctl==2.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
zipp==3.4.1
